+++ core-ng/src/main/java/core/framework/module/APIClientConfig.java
@@ -1,15 +1,15 @@
 package core.framework.module;
 
-import core.framework.impl.web.service.WebServiceClient;
+import core.framework.impl.web.service.WebServiceClientProxy;
 import core.framework.web.service.WebServiceClientInterceptor;
 
 /**
  * @author neo
  */
 public final class APIClientConfig {     // returned from APIConfig, must be public to be called
-    private final WebServiceClient client;
+    private final WebServiceClientProxy client;
 
-    APIClientConfig(WebServiceClient client) {
+    APIClientConfig(WebServiceClientProxy client) {
         this.client = client;
     }